Java第20回Read処理のServiceとController部分を実装
