+++ src/main/java/raisetech/StudentManagement/StudentManagementApplication.java
@@ -6,25 +6,15 @@ import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RestController;
+import raisetech.StudentManagement.data.Student;
+import raisetech.StudentManagement.data.StudentsCourses;
+import raisetech.StudentManagement.repositry.StudentRepository;
 
 @SpringBootApplication
-@RestController
-public class StudentManagementApplication {
 
-  @Autowired
-  private StudentRepository repository;
+public class StudentManagementApplication {
 
   public static void main(String[] args) {
     SpringApplication.run(StudentManagementApplication.class, args);
-  }
-
-  @GetMapping("/studentList")
-  public List<Student> getStudentList() {
-    return repository.search();
-  }
-
-  @GetMapping("/studentsCourseList")
-  public List<StudentsCourses> getStudentsCoursesList() {
-    return repository.searchStudentsCourses();
   }
 }
