+++ src/main/java/raisetech/StudentManagement/service/StudentService.java
@@ -1,0 +1,31 @@
+package raisetech.StudentManagement.service;
+
+import java.util.List;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Service;
+import raisetech.StudentManagement.data.Student;
+import raisetech.StudentManagement.data.StudentsCourses;
+import raisetech.StudentManagement.repositry.StudentRepository;
+
+@Service
+public class StudentService {
+
+  private StudentRepository repository;
+
+  @Autowired
+  public StudentService(StudentRepository repository) {
+    this.repository = repository;
+  }
+
+  public List<Student> searchStudentList() {
+    /* 検索処理
+    repository.search();*/
+    // 絞り込みをして、年齢が30代の人のみ抽出する。抽出したリストをControllerに返す。
+    return repository.search();
+  }
+
+  public List<StudentsCourses> searchStudentsCoursesList() {
+    // 絞り込み検索で、「Java」コウス情報のみを抽出する。抽出したリストをControllerに返す。
+    return repository.searchStudentsCourses();
+  }
+}
